+++ Docs/releases.md
@@ -5,6 +5,7 @@
  * Add `FluxorOptions.ScanTypes` to allow scanning of specified classes. ([#214](https://github.com/mrpmorris/Fluxor/issues/214))
  * Make `FluxorComponent` and `FluxorLayout` abstract. ([#217](https://github.com/mrpmorris/Fluxor/issues/217))
  * Add `ForceLoad` property to `GoAction`. ([#178](https://github.com/mrpmorris/Fluxor/issues/178))
+ * Only call `Dispose(bool disposing)` if not already exposed. ([#222](https://github.com/mrpmorris/Fluxor/issues/222))
 
 ## New in 4.1
  * Allow custom control over JSON serialisation in Redux Dev Tools - see